+++ packages/desktop/app/views/LabRequestListingView.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { TopBar, PageContainer, LabRequestsSearchBar, ContentPane } from '../components';
-import { LabRequestsTable } from '../components/LabRequestsTable';
+import { LabRequestsTable } from './LabRequestsTable';
 
 export const LabRequestListingView = React.memo(() => (
   <PageContainer>
