+++ packages/desktop/app/views/patients/EncounterLabRequestsTable.js
@@ -1,0 +1,49 @@
+import React from 'react';
+import { useDispatch } from 'react-redux';
+import { useParams } from 'react-router-dom';
+import { push } from 'connected-react-router';
+import { DataFetchingTable } from '../../components';
+import { reloadPatient } from '../../store/patient';
+import { useLabRequest } from '../../contexts/LabRequest';
+import {
+  getRequestedBy,
+  getStatus,
+  getRequestType,
+  getPriority,
+  getDateTime,
+  getRequestId,
+} from '../../utils/lab';
+
+const columns = [
+  { key: 'requestId', title: 'Test ID', sortable: false, accessor: getRequestId },
+  { key: 'category.name', title: 'Test category', accessor: getRequestType },
+  { key: 'requestedDate', title: 'Requested at time', accessor: getDateTime },
+  { key: 'displayName', title: 'Requested by', accessor: getRequestedBy, sortable: false },
+  { key: 'priority.name', title: 'Priority', accessor: getPriority },
+  { key: 'status', title: 'Status', accessor: getStatus, maxWidth: 200 },
+];
+
+export const EncounterLabRequestsTable = React.memo(({ encounterId }) => {
+  const { patientId, category } = useParams();
+  const dispatch = useDispatch();
+  const { loadLabRequest, searchParameters } = useLabRequest();
+
+  const selectLab = async lab => {
+    if (lab.patientId) await dispatch(reloadPatient(lab.patientId));
+    await loadLabRequest(lab.id);
+    dispatch(
+      push(`/patients/${category}/${patientId}/encounter/${encounterId}/lab-request/${lab.id}`),
+    );
+  };
+
+  return (
+    <DataFetchingTable
+      endpoint={`encounter/${encounterId}/labRequests`}
+      columns={columns}
+      noDataMessage="No lab requests found"
+      onRowClick={selectLab}
+      fetchOptions={searchParameters}
+      initialSort={{ order: 'desc', orderBy: 'requestedDate' }}
+    />
+  );
+});
